pyproject.toml
[tool.poetry]
name = "s3-artifact"
version = "0.1.0"
description = ""
authors = ["Marcel Arns <marcel.arns@moneymeets.com>"]

[tool.poetry.dependencies]
python = "~3.12"

dataclasses-json = "*"
pyyaml = "*"

[tool.poetry.dev-dependencies]
ruff = "*"

pytest = "*"
pytest-cov = "*"

[tool.coverage.run]
branch = true
source = ["."]
omit = ["**/tests/*.py"]

[tool.poetry.scripts]
s3-artifact = 's3_artifact.action:main'

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
